requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements-to-freeze.txt
#
deap==1.0.2.post2
decorator==4.0.10
dill==0.2.5
mpmath==0.19
networkx==1.11
numpy==1.11.2
pyyaml==3.12
scipy==0.18.1
sympy==1.0
toolz==0.8.1
